+++ src/main/java/MVC/DAO/clienteDAOClase.java
@@ -46,7 +46,7 @@ public class clienteDAOClase implements clienteDAO{
 
     @Override
     @Transactional
-    public Cliente getCliente(long id) {
+    public Cliente getCliente(int id) {
         Session miSesion=sessionFactory.openSession();
         //Obterner informacion del cliente selecionado
         Cliente elCliente=miSesion.get(Cliente.class,id);
@@ -55,7 +55,7 @@ public class clienteDAOClase implements clienteDAO{
 
     @Override
     @Transactional
-    public void eliminarCliente(long id) {
+    public void eliminarCliente(int id) {
         Session miSesion=sessionFactory.openSession();
         Transaction tx = miSesion.beginTransaction();
         //Borrar el cliente de la base de datos usando su id correspondiente
